requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
motor==3.7.0
selenium==4.8.0
beautifulsoup4==4.11.1
httpx==0.23.0
python-dotenv==0.20.0
requests==2.28.1